refactor(git-frontend): Modularize UI components and state management
- **Isolate Code**: Moved all Git-related Vue components, API calls, and utility functions into a dedicated `client/git-integration/` directory.
- **Refactor State**: Decomposed the monolithic `gitStore.js` into five specialized Pinia stores (`panelUi`, `status`, `history`, `actions`, `log`) for improved maintainability.
- **Re-integrate Components**: Updated all Git components to use the new modular stores, completing the frontend architecture overhaul.
