+++ client/git-integration/components/tabs/GitHistoryTab.vue
@@ -1,18 +1,17 @@
-<!-- client/components/GitPanel/GitHistoryTab.vue -->
+<!-- client/git-integration/components/tabs/GitHistoryTab.vue -->
 <template>
   <div class="flex h-full flex-col p-2">
     <div
-      v-if="store.gitLog.length === 0 && !store.isActionLoading"
+      v-if="historyStore.gitLog.length === 0 && !historyStore.isLoading"
       class="py-4 text-center text-theme-text-muted"
     >
       No commit history found.
     </div>
-
     <div v-else class="min-h-0 flex-grow overflow-y-auto">
-      <div v-for="commit in store.gitLog" :key="commit.hash">
+      <div v-for="commit in historyStore.gitLog" :key="commit.hash">
         <!-- Commit Header (Clickable) -->
         <div
-          @click="store.toggleCommitExpansion(commit.hash)"
+          @click="historyStore.toggleCommitExpansion(commit.hash)"
           class="flex cursor-pointer items-center justify-between rounded p-2 hover:bg-theme-border"
         >
           <div class="flex-grow pr-2">
@@ -31,7 +30,7 @@
             <SvgIcon
               type="mdi"
               :path="
-                store.expandedCommit === commit.hash
+                historyStore.expandedCommit === commit.hash
                   ? mdilChevronUp
                   : mdilChevronDown
               "
@@ -40,22 +39,24 @@
             />
           </div>
         </div>
-
         <!-- Expanded File List -->
         <div
-          v-if="store.expandedCommit === commit.hash"
+          v-if="historyStore.expandedCommit === commit.hash"
           class="ml-4 border-l-2 border-theme-border pl-2"
         >
           <div
-            v-if="store.isActionLoading && !store.commitFilesCache[commit.hash]"
+            v-if="
+              historyStore.isFilesLoading &&
+              !historyStore.commitFilesCache[commit.hash]
+            "
             class="py-2 text-sm text-theme-text-muted"
           >
             Loading files...
           </div>
           <div
             v-else-if="
-              store.commitFilesCache[commit.hash] &&
-              store.commitFilesCache[commit.hash].length === 0
+              historyStore.commitFilesCache[commit.hash] &&
+              historyStore.commitFilesCache[commit.hash].length === 0
             "
             class="py-2 text-sm text-theme-text-muted"
           >
@@ -63,7 +64,7 @@
           </div>
           <div v-else>
             <div
-              v-for="file in store.commitFilesCache[commit.hash]"
+              v-for="file in historyStore.commitFilesCache[commit.hash]"
               :key="file.path"
               class="flex items-center justify-between py-1 text-sm"
             >
@@ -71,7 +72,7 @@
               <div class="ml-2 flex flex-shrink-0 items-center">
                 <button
                   v-if="file.path.endsWith('.md')"
-                  @click="store.openNoteInEditor(file.path)"
+                  @click="historyStore.openNoteInEditor(file.path)"
                   class="p-1 text-theme-text-muted hover:text-theme-text"
                   title="Open File"
                 >
@@ -90,12 +91,11 @@
     </div>
   </div>
 </template>
-
 <script setup>
-import { useGitStore } from "../../gitStore";
+import { useHistoryStore } from "../../stores/historyStore";
 import { getCommitFileStatusClass } from "../../gitUtils";
 import SvgIcon from "@jamescoyle/vue-icon";
 import { mdilFile, mdilChevronDown, mdilChevronUp } from "@mdi/light-js";
 
-const store = useGitStore();
+const historyStore = useHistoryStore();
 </script>
