+++ client/git-integration/gitUtils.js
