+++ client/git-integration/stores/statusStore.js
@@ -1,0 +1,121 @@
+// client/git-integration/stores/statusStore.js
+import { defineStore } from "pinia";
+import { ref, computed } from "vue";
+import { useToast } from "primevue/usetoast";
+import { useRouter } from "vue-router";
+import * as gitApi from "../gitApi";
+
+/**
+ * Manages the state related to `git status`, including staged/unstaged files,
+ * the commit message, and the summary for the status indicator.
+ */
+export const useStatusStore = defineStore("git-status", () => {
+  // -- DEPS --
+  const toast = useToast();
+  const router = useRouter();
+
+  // -- STATE --
+  const gitStatus = ref({ files: [] });
+  const commitMessage = ref("");
+  const isLoading = ref(true); // For detailed status
+  let summaryPollInterval = null;
+
+  // For GitStatusIndicator
+  const isLoadingSummary = ref(true);
+  const summaryError = ref(null);
+  const branchName = ref("");
+  const filesChangedCount = ref(0);
+
+  // -- GETTERS --
+  const stagedFiles = computed(() =>
+    gitStatus.value.files.filter(
+      (f) => f.index_status !== " " && f.index_status !== "?",
+    ),
+  );
+
+  const unstagedFiles = computed(() =>
+    gitStatus.value.files.filter((f) => f.work_tree_status !== " "),
+  );
+
+  const tooltipText = computed(() => {
+    if (isLoadingSummary.value) return "Loading Git status...";
+    if (summaryError.value) return `Error: ${summaryError.value}`;
+    if (filesChangedCount.value > 0)
+      return `${filesChangedCount.value} changes detected on branch '${branchName.value}'. Click to view.`;
+    return `On branch '${branchName.value}'. Synced.`;
+  });
+
+  // -- ACTIONS --
+  async function fetchStatus() {
+    isLoading.value = true;
+    try {
+      gitStatus.value = await gitApi.getGitStatus();
+    } catch (err) {
+      toast.add({
+        severity: "error",
+        summary: "Error Fetching Status",
+        detail: err.message,
+        life: 3000,
+      });
+      gitStatus.value = { files: [] }; // Reset on error
+    } finally {
+      isLoading.value = false;
+    }
+  }
+
+  async function fetchStatusSummary() {
+    try {
+      const summary = await gitApi.getGitStatusSummary();
+      branchName.value = summary.current_branch || "N/A";
+      filesChangedCount.value = summary.files_changed_count;
+      summaryError.value = null;
+    } catch (err) {
+      summaryError.value = err.response?.data?.detail || "Failed to connect";
+      console.error("StatusStore: Failed to fetch git status summary:", err);
+    } finally {
+      isLoadingSummary.value = false;
+    }
+  }
+
+  function openNoteInEditor(path) {
+    const title = path.replace(/\.md$/, "");
+    router.push({ name: "note", params: { title } });
+  }
+
+  function clearCommitMessage() {
+    commitMessage.value = "";
+  }
+
+  function initialize() {
+    isLoadingSummary.value = true;
+    fetchStatus();
+    fetchStatusSummary();
+    summaryPollInterval = setInterval(fetchStatusSummary, 30000);
+  }
+
+  function cleanup() {
+    if (summaryPollInterval) clearInterval(summaryPollInterval);
+  }
+
+  return {
+    // State
+    gitStatus,
+    commitMessage,
+    isLoading,
+    isLoadingSummary,
+    summaryError,
+    branchName,
+    filesChangedCount,
+    // Getters
+    stagedFiles,
+    unstagedFiles,
+    tooltipText,
+    // Actions
+    fetchStatus,
+    fetchStatusSummary,
+    openNoteInEditor,
+    clearCommitMessage,
+    initialize,
+    cleanup,
+  };
+});
